code is more organized
