+++ .gitignore
@@ -8,5 +8,9 @@ yolov5su.pt
 yolov8s.pt
 /.vscode
 /__pycache__/
+/app/
 tracker.py
 tracking.py
+test.py
+/test_gallery/
+video.mp4
